 requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1
onnxruntime>=1.15.0
opencv-python>=4.8.0
mediapipe>=0.10.0
gradio>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
